HHH-15165 Add abstract GIS base types to reduce code duplication
